Switches from deprecated DiscoveryClient service ID mechanism to Registration
